completed day6
this was a challenging day, I initially wrote a conventional implementation that goes through every fish and updates it, however this way to slow to deal with the large amounts of fishes at the end, I did several different iterations of the code with optimisation but simply to slow for the operations needed.

I also tried to model the predicted amount of fish using a exponential growth model but that didn't quit work out.

in the end I was collaborating with my brother who had a clever and super optimised solution for this that could handle any amount of days up to the maximum allowed int value.